feat : Add context in order to log the node path
We can now add a context to be able to log in its value all the resolved node the engine went through during the resolve process.
adding a context to the tree :
```golang
t.WithContext(context.Background())
```
and after the resolve geting the slice of string generated.
```golang
s := dtree.GetNodePathFromContext(t.Context())
```